ml-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
pydantic==2.5.0